backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
pydantic>=2.9.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
aiofiles>=23.2.1
openai>=1.3.5
numpy>=1.26.2
websockets>=12.0
python-socketio>=5.10.0
resend>=0.6.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

